weather_app/requirements.txt
flet==0.28.3
httpx>=0.25.0
python-dotenv>=1.0.0
EOF